Split Modifier file
