+++ Content/Data/Modifications_Explosive.cs
@@ -1,0 +1,70 @@
+﻿using TC2.Base.Components;
+
+namespace TC2.Base
+{
+	public sealed partial class ModInstance
+	{
+		private static void RegisterExplosiveModifications(ref List<Modification.Definition> definitions)
+		{
+			definitions.Add(Modification.Definition.New<Explosive.Data>
+			(
+				identifier: "explosive.directed",
+				category: "Explosives",
+				name: "Directed Explosion",
+				description: "Focuses the explosion into a smaller area.",
+
+				apply_0: static (ref Modification.Context context, ref Explosive.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					//var radius_log = MathF.Log2(data.radius);
+
+					data.radius *= 0.75f;
+					data.power *= 1.25f;
+					data.damage_terrain *= 1.25f;
+					data.damage_entity *= 1.25f;
+				}
+			));
+
+			definitions.Add(Modification.Definition.New<Explosive.Data>
+			(
+				identifier: "explosive.nitroglycerine",
+				category: "Explosives",
+				name: "Nitroglycerine Filler",
+				description: "Replaces filler with nitroglycerine.",
+
+				apply_0: static (ref Modification.Context context, ref Explosive.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					var material_nitroglycerine_handle = new Material.Handle("nitroglycerine");
+					ref var material_nitroglycerine = ref material_nitroglycerine_handle.GetDefinition();
+
+					var has_any = false;
+
+					foreach (ref var requirement in context.requirements_new)
+					{
+						if (requirement.type == Crafting.Requirement.Type.Resource)
+						{
+							ref var material = ref requirement.material.GetDefinition();
+							if (requirement.material.id != material_nitroglycerine_handle.id && material.flags.HasAny(Material.Flags.Explosive))
+							{
+								var amount_new = (requirement.amount * material.mass_per_unit) / material_nitroglycerine.mass_per_unit;
+
+								requirement.amount = amount_new;
+								requirement.material = material_nitroglycerine_handle;
+
+								has_any = true;
+							}
+						}
+					}
+
+					if (has_any)
+					{
+						data.radius += MathF.Sqrt(data.radius * 1.50f);
+						data.power += MathF.Sqrt(data.power * 2.50f);
+						data.damage_terrain += MathF.Pow(data.damage_terrain * 3.50f, 0.75f);
+						data.damage_entity += MathF.Pow(data.damage_entity * 2.50f, 0.75f);
+					}
+				}
+			));
+		}
+	}
+}
+
