+++ Content/Data/Modifications_Fuse.cs
@@ -1,0 +1,46 @@
+﻿using TC2.Base.Components;
+
+namespace TC2.Base
+{
+	public sealed partial class ModInstance
+	{
+		private static void RegisterFuseModifications(ref List<Modification.Definition> definitions)
+		{
+			definitions.Add(Modification.Definition.New<Fuse.Data>
+			(
+				identifier: "fuse.length",
+				category: "Explosives",
+				name: "Fuse Length",
+				description: "Modifies fuse's length.",
+
+#if CLIENT
+				draw_editor: static (ref Modification.Context context, in Fuse.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var value = ref handle.GetData<float>();
+					return GUI.SliderFloat("##stuff", ref value, 0.50f, 10.00f, "%.2f");
+				},
+#endif
+
+				apply_0: static (ref Modification.Context context, ref Fuse.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var value = ref handle.GetData<float>();
+					data.time = value;
+				}
+			));
+
+			definitions.Add(Modification.Definition.New<Fuse.Data>
+			(
+				identifier: "fuse.inextinguishable",
+				category: "Explosives",
+				name: "Inextinguishable Fuse",
+				description: "Makes the fuse impossible to be extinguished.",
+
+				apply_0: static (ref Modification.Context context, ref Fuse.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					data.failure_chance = 0.00f;
+				}
+			));
+		}
+	}
+}
+
