+++ Content/Data/Modifications_Health.cs
@@ -1,0 +1,133 @@
+﻿using TC2.Base.Components;
+
+namespace TC2.Base
+{
+	public sealed partial class ModInstance
+	{
+		private static void RegisterHealthModifications(ref List<Modification.Definition> definitions)
+		{
+			//definitions.Add(Modification.Definition.New<Health.Data>
+			//(
+			//	identifier: "health.reinforced_structure",
+			//	name: "Reinforced Structure",
+			//	description: "Increases durability.",
+
+			//	can_add: static (ref Modification.Context context, in Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+			//	{
+			//		var count = 0;
+			//		for (int i = 0; i < modifications.Length; i++)
+			//		{
+			//			if (modifications[i].id == handle.id) count++;
+			//		}
+			//		return count < 1;
+			//	},
+
+			//	apply_0: static (ref Modification.Context context, ref Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+			//	{
+			//		data.max *= 1.20f;
+			//	},
+
+			//	apply_1: static (ref Modification.Context context, ref Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+			//	{
+			//		foreach (ref var requirement in context.requirements_new)
+			//		{
+			//			if (requirement.type == Crafting.Requirement.Type.Work)
+			//			{
+			//				requirement.amount *= 1.20f;
+			//			}
+			//		}
+			//	}
+			//));
+
+			definitions.Add(Modification.Definition.New<Health.Data>
+			(
+				identifier: "health.varnish",
+				category: "Health",
+				name: "Varnish",
+				description: "Applies varnish on item's surface, improving its durability.",
+
+				can_add: static (ref Modification.Context context, in Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					return context.requirements_new.Has(Crafting.Requirement.Resource("wood", 0.00f)) && !modifications.HasModification(handle);
+				},
+
+				apply_1: static (ref Modification.Context context, ref Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					var amount = 0.00f;
+					foreach (ref var requirement in context.requirements_new)
+					{
+						if (requirement.type == Crafting.Requirement.Type.Resource)
+						{
+							ref var material = ref requirement.material.GetDefinition();
+							if (material.type == Material.Type.Wood)
+							{
+								amount += requirement.amount;
+							}
+						}
+					}
+
+					var total_amount = 3.00f + (amount * 0.10f);
+					context.requirements_new.Add(Crafting.Requirement.Resource("resin", total_amount));
+
+					data.max += total_amount * 85.00f;
+					data.dt += 50.00f;
+
+					ref var body = ref context.GetComponent<Body.Data>();
+					if (!body.IsNull())
+					{
+						ref var material = ref Material.GetMaterial("resin");
+						body.mass_extra += total_amount * material.mass_per_unit * 0.30f;
+					}
+				}
+			));
+
+			definitions.Add(Modification.Definition.New<Health.Data>
+			(
+				identifier: "health.smirgl_structure",
+				category: "Health",
+				name: "Smirgl-Reinforced Structure",
+				description: "Replaces entire structure with smirgl, greatly increasing durability.",
+
+				can_add: static (ref Modification.Context context, in Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					return !modifications.HasModification(handle);
+				},
+
+				apply_1: static (ref Modification.Context context, ref Health.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					var ingot_amount = 0.00f;
+					foreach (ref var requirement in context.requirements_new)
+					{
+						if (requirement.type == Crafting.Requirement.Type.Work)
+						{
+							requirement.amount *= 2.00f;
+							requirement.difficulty += 4.00f;
+						}
+						else if (requirement.type == Crafting.Requirement.Type.Resource)
+						{
+							ref var material = ref requirement.material.GetDefinition();
+							if (material.flags.HasAll(Material.Flags.Ingot))
+							{
+								ingot_amount += requirement.amount;
+								requirement = default;
+							}
+						}
+					}
+
+					var total_amount = 3.00f + (ingot_amount * 0.30f);
+					context.requirements_new.Add(Crafting.Requirement.Resource("smirgl_ingot", total_amount));
+
+					data.max += total_amount * 2500.00f;
+
+					ref var body = ref context.GetComponent<Body.Data>();
+					if (!body.IsNull())
+					{
+						ref var material = ref Material.GetMaterial("smirgl_ingot");
+						body.mass_extra += total_amount * material.mass_per_unit * 0.70f;
+					}
+				}
+			));
+		}
+	}
+}
+
