+++ Content/Data/Modifications_Overheat.cs
@@ -1,0 +1,101 @@
+﻿using TC2.Base.Components;
+
+namespace TC2.Base
+{
+	public sealed partial class ModInstance
+	{
+		private static void RegisterOverheatModifications(ref List<Modification.Definition> definitions)
+		{
+			definitions.Add(Modification.Definition.New<Overheat.Data>
+			(
+				identifier: "overheat.coolant",
+				category: "Cooling",
+				name: "Water-Cooled",
+				description: "Increases cooling rate.",
+
+				validate: static (ref Modification.Context context, in Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var amount = ref handle.GetData<int>();
+					amount = Maths.Clamp(amount, 1, 40);
+
+					return true;
+				},
+
+#if CLIENT
+				draw_editor: static (ref Modification.Context context, in Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var amount = ref handle.GetData<int>();
+					return GUI.SliderInt("##stuff", ref amount, 1, 40, "%d");
+				},
+#endif
+
+				apply_0: static (ref Modification.Context context, ref Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var amount = ref handle.GetData<int>();
+					data.cool_rate += amount * 4.00f;
+				},
+
+				apply_1: static (ref Modification.Context context, ref Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					ref var amount = ref handle.GetData<int>();
+
+					context.requirements_new.Add(Crafting.Requirement.Resource("water", amount));
+
+					ref var body = ref context.GetComponent<Body.Data>();
+					if (!body.IsNull())
+					{
+						ref var material = ref Material.GetMaterial("water");
+						body.mass_extra += amount * material.mass_per_unit;
+					}
+				}
+			));
+
+			definitions.Add(Modification.Definition.New<Overheat.Data>
+			(
+				identifier: "overheat.heat_resistant",
+				category: "Cooling",
+				name: "Heat-Resistant Components",
+				description: "Dramatically increases maximum operating temperature at cost of extra weight and reduced cooling rate.",
+
+				can_add: static (ref Modification.Context context, in Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					return !modifications.HasModification(handle);
+				},
+
+				apply_0: static (ref Modification.Context context, ref Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					data.cool_rate *= 0.80f;
+					data.heat_critical += 500.00f;
+				},
+
+				apply_1: static (ref Modification.Context context, ref Overheat.Data data, ref Modification.Handle handle, Span<Modification.Handle> modifications) =>
+				{
+					var amount = 0.00f;
+
+					foreach (ref readonly var requirement in context.requirements_old)
+					{
+						if (requirement.type == Crafting.Requirement.Type.Resource)
+						{
+							ref var material = ref requirement.material.GetDefinition();
+							if (material.flags.HasAny(Material.Flags.Manufactured | Material.Flags.Metal))
+							{
+								amount += requirement.amount * 0.20f;
+							}
+						}
+					}
+
+					amount = MathF.Ceiling(amount);
+					context.requirements_new.Add(Crafting.Requirement.Resource("smirgl_ingot", amount));
+
+					ref var body = ref context.GetComponent<Body.Data>();
+					if (!body.IsNull())
+					{
+						ref var material = ref Material.GetMaterial("smirgl_ingot");
+						body.mass_extra += amount * material.mass_per_unit;
+					}
+				}
+			));
+		}
+	}
+}
+
