+++ content/Mod.cs
@@ -5,6 +5,15 @@ namespace TC2.Base
 	{
 		protected override void OnRegister(ModContext context)
 		{
+			Modification.OnInitialize += RegisterGunModifications;
+			Modification.OnInitialize += RegisterHealthModifications;
+			Modification.OnInitialize += RegisterBodyModifications;
+			Modification.OnInitialize += RegisterFuseModifications;
+			Modification.OnInitialize += RegisterExplosiveModifications;
+			Modification.OnInitialize += RegisterMeleeModifications;
+			Modification.OnInitialize += RegisterDrillModifications;
+			Modification.OnInitialize += RegisterMedkitModifications;
+			Modification.OnInitialize += RegisterOverheatModifications;
 			Modification.OnInitialize += RegisterModifications;
 		}
